Implement size checks for sessions and session attributes.
Recently we experienced a production issue caused by some automated 3rd
party bot/script that resulted in a session attribute growing unbounded.
This lead to network saturation as the manager dutifully saved and
loaded the ever-growing session and caused a denial-of-service against
valid sessions.
Although the root cause was fixed, we felt that having an additional
layer that limited session size would be A Good Thing (tm).

README.md contains details on the new configuration values. If you're
considering using this, you may want to write a simple script that
iterates over all the current persisted sessions in redis and dumps
their size to ensure you're not setting this to a too-small value. Or
just use `allowOversizedSessions="true"`